[fixup] mktime: add casts

+++ CryptoPkg/Library/BaseCryptLib/SysCall/TimerWrapper.c
@@ -148,12 +148,12 @@ mktime (
   )
 {
   EFI_TIME  Time = {
-    .Year     = t->tm_year,
-    .Month    = t->tm_mon,
-    .Day      = t->tm_mday,
-    .Hour     = t->tm_hour,
-    .Minute   = t->tm_min,
-    .Second   = t->tm_sec,
+    .Year     = (UINT16)t->tm_year,
+    .Month    = (UINT8)t->tm_mon,
+    .Day      = (UINT8)t->tm_mday,
+    .Hour     = (UINT8)t->tm_hour,
+    .Minute   = (UINT8)t->tm_min,
+    .Second   = (UINT8)t->tm_sec,
     .TimeZone = EFI_UNSPECIFIED_TIMEZONE,
   };
